+++ lib/bike_brigade/mailchimp_api/http.ex
@@ -6,7 +6,7 @@ defmodule BikeBrigade.MailchimpApi.Http do
   @count 100
 
   @impl MailchimpApi
-  def get_list(list_id, last_changed \\ nil) do
+  def get_list(list_id, opted_in \\ nil) do
     with {:ok, account} <- Mailchimp.Account.get(),
          {:ok, list} <- Mailchimp.Account.get_list(account, list_id) do
       # Infinite sequence of offsets 0,100,200,...
@@ -19,7 +19,7 @@ defmodule BikeBrigade.MailchimpApi.Http do
                  offset: offset,
                  fields:
                    "members.email_address,members.id,members.status,members.merge_fields,members.timestamp_opt",
-                 since_last_changed: last_changed
+                 since_timestamp_opt: opted_in
                }) do
             {:ok, []} -> {:halt, :ok}
             {:ok, members} -> {members, :ok}
@@ -28,6 +28,17 @@ defmodule BikeBrigade.MailchimpApi.Http do
         end)
 
       {status, members}
+    end
+  end
+
+  @impl MailchimpApi
+  def update_member_fields(list_id, email, fields) do
+    with {:ok, account} <- Mailchimp.Account.get(),
+         {:ok, list} <- Mailchimp.Account.get_list(account, list_id),
+         {:ok, member} <- Mailchimp.List.get_member(list, email) do
+      member
+      |> Map.update(:merge_fields, %{}, &Map.merge(&1, fields))
+      |> Mailchimp.Member.update()
     end
   end
 end